refactor: gameScore 관련 패키지 네이밍 변경
