+++ backend/src/main/java/com/now/naaga/game/domain/gamescore/GameScorePolicy.java
@@ -1,10 +1,10 @@
-package com.now.naaga.game.domain.scorestrategy;
+package com.now.naaga.game.domain.gamescore;
 
 import com.now.naaga.game.domain.Game;
 import com.now.naaga.game.domain.ResultType;
 import com.now.naaga.score.domain.Score;
 
-public interface ScorePolicy {
+public interface GameScorePolicy {
 
     Score calculate(final Game game);
